+++ packages/block-library/src/social-link/edit.native.js
@@ -158,7 +158,7 @@ const SocialLinkEdit = ( {
 		  );
 
 	return (
-		<View style={ styles.container }>
+		<View>
 			{ isSelected && (
 				<>
 					<BlockControls>